Use models to determine tool
In the previous version, the choose of model used to use the first
model but it will waste a lot of choices of models so that in this
commit, I make the model to determine the tool.

For example, now, the available models are UTC2000, UTC3000, UTC5000,
RAPID, MAXUM, MAXUM-PLUS and the tools are A0801XXX, A0803XXX. In the
previous version, the program tends to use A0801 so that the models will
only be UTC2000 and UTC3000. Actually, if it chooses A0803, the result
will be definitely better because it has more models to use such as
UTC5000, RAPID, MAXUM, MAXUM-PLUS.