Attempt to optimize machine distributing method
In this commit, I modified the function _chooseMachinesForAGroup
attempting to optimize the distributing method. The way I used is
determining minimum set of machines first and if there still are
remaining resources, the function would look around the machine set and
add the available machines until running out of resources.